% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cep-cnd-conditions.R
\name{cep-cnd-conditions}
\alias{cep-cnd-conditions}
\alias{cep:bad_argument}
\alias{bad_argument}
\alias{cep:bad_argument/error}
\title{Conditions for \code{cep}}
\description{
Conditions for \code{cep}
}
\details{
The following conditions are defined in the \code{{cep}} package.
}
\section{\code{\link[=cnd-package]{\{cnd\}}}}{

These conditions are made with the \code{{cnd}} package though the use of
\code{\link[cnd:condition]{cnd::condition()}}.
}

\section{\code{{cep}} conditions}{


\subsection{\code{cep:bad_argument/error}}{
\describe{
\item{package}{\code{{cep}}}
\item{class}{\code{cep:bad_argument}}
\item{type}{\strong{error}}
}
x should be \code{TRUE}
}
}

\seealso{
\link[cnd:cnd-package]{cnd::cnd-package} \link[cnd:condition]{cnd::condition}
}
\keyword{internal}
