% Generated with cnd::cnd_document("cep")
\name{cep-conditions}
\alias{cep-conditions}
\alias{cep:bad_argument}
\title{Conditions for \code{{cep}}}
\description{
  The following conditions are defined in the cep package.
  For more information on \code{conditions} see
  \code{\link[cnd:conditions]{cnd::conditions()}}
}

\section{\code{cep:bad_argument}}{
\describe{
  \item{class}{cep:bad_argument}
  \item{type}{error}
  \item{package}{cep}
}

see: \code{\link[cnd:example_function]{example_function()}}
}
