% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{example_function}
\alias{example_function}
\title{An example function}
\usage{
example_function(x)
}
\description{
An example function
}
\section{Conditions}{


\describe{

\item{\code{cep:bad_argument}}{

}

}
\seealso{
\code{\link{cep-conditions}}
}
}

