% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cep-cnd-conditions.R
\docType{package}
\name{cep-cnd-conditions}
\alias{cep}
\alias{cep-package}
\alias{cep-cnd-conditions}
\alias{cep:bad_argument}
\alias{bad_argument}
\alias{cep:bad_argument/error}
\title{Conditions for \code{cep}}
\description{
Example package used for testing cep
}
\details{
The following conditions are defined in the \code{{cep}} package.
}
\section{\code{\link[cnd}{
cnd-package]{\{cnd\}}}:
These conditions are made with the \code{{cnd}} package though the use of
\code{\link[cnd:condition]{cnd::condition()}}.
}

\section{\code{{cep}} conditions}{


\subsection{\code{cep:bad_argument/error}}{
\describe{
\item{package}{\code{{cep}}}
\item{class}{\code{cep:bad_argument}}
\item{type}{\strong{error}}
}

}
}

\seealso{
\link[cnd:cnd-package]{cnd::cnd-package} \link[cnd:condition]{cnd::condition}
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
