% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{example_function}
\alias{example_function}
\alias{bad_argument}
\title{An example function}
\usage{
example_function(x)

bad_argument(..., .call = getOption("cnd.call", TRUE))
}
\description{
An example function
}
\section{Conditions}{


Conditions are generated through the \code{\link[cnd:cnd-package]{\{cnd\}}} package.
The following conditions are associated with this function:

\describe{

}

For more conditions, see:
}

