% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{example_function}
\alias{example_function}
\title{An example function}
\usage{
example_function(x)
}
\description{
An example function
}
\section{Conditions}{


Conditions are generated through the \code{\link[cnd:cnd-package]{\{cnd\}}} package.
The following conditions are associated with this function:

\describe{

\item{\code{\link[=cep-conditions]{cep:bad_argument/error}}}{

}

}

For more conditions, see: \link{cep-conditions}
}

